Moved the Web-Crawler to the @ckeditor/ckeditor5-dev-docs package.
