+++ package.json
@@ -108,15 +108,12 @@
     "minimist": "^1.2.5",
     "mkdirp": "^1.0.4",
     "nyc": "^15.0.1",
-    "ora": "^5.2.0",
     "popper": "^1.0.1",
     "postcss-loader": "^3.0.0",
     "progress-bar-webpack-plugin": "^2.1.0",
-    "puppeteer": "^8.0.0",
     "raw-loader": "^4.0.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "strip-ansi": "^6.0.0",
     "style-loader": "^1.2.1",
     "stylelint": "^13.5.0",
     "stylelint-config-ckeditor5": "^2.0.0",
